feat: main functional

+++ src/utils/Api.js
@@ -1,3 +1,5 @@
+import Card from '../components/Card';
+
 class Api {
   constructor(data) {
     this._baseUrl = data.serverUrl;
@@ -69,21 +71,37 @@ class Api {
     return this._requestResult(res);
   }
 
-  async putLike(cardId) {
-    const res = await fetch(`${this._baseUrl}cards/${cardId}/likes`, {
-      method: 'PUT',
-      headers: this._headers,
-    });
-    return this._requestResult(res);
+  async changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      const res = await fetch(`${this._baseUrl}cards/${cardId}/likes`, {
+        method: 'PUT',
+        headers: this._headers,
+      });
+      return this._requestResult(res);
+    } else {
+      const res = await fetch(`${this._baseUrl}cards/${cardId}/likes`, {
+        method: 'DELETE',
+        headers: this._headers,
+      });
+      return this._requestResult(res);
+    }
   }
 
-  async removeLike(cardId) {
-    const res = await fetch(`${this._baseUrl}cards/${cardId}/likes`, {
-      method: 'DELETE',
-      headers: this._headers,
-    });
-    return this._requestResult(res);
-  }
+  // async putLike(cardId) {
+  //   const res = await fetch(`${this._baseUrl}cards/${cardId}/likes`, {
+  //     method: 'PUT',
+  //     headers: this._headers,
+  //   });
+  //   return this._requestResult(res);
+  // }
+
+  // async removeLike(cardId) {
+  //   const res = await fetch(`${this._baseUrl}cards/${cardId}/likes`, {
+  //     method: 'DELETE',
+  //     headers: this._headers,
+  //   });
+  //   return this._requestResult(res);
+  // }
 }
 
 export const api = new Api({
